finished with the header/nav/hero section
added visibility property to the nav wrap so that I could apply a transition on the background color so it will work on mobile.

I have had problems with css animations only running once on mobile and not repeating the animation. I found that transition property works best because it will work over and over again